Revamp the test_load_composable_nodes test. (#403)
* Revamp the test_load_composable_nodes test.

In particular, it made little sense to do an rclpy.init
in a class that inherited from rclpy.node.Node.  Instead,
flip this around and make the mock_component_container
responsible for creating the context, doing rclpy.init,
creating the node, spinning the executor, and tearing it
all down.  Note that this uses the new rclpy.init
context manager as well, which reduces the amount of code
we need here.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>